requirements.txt
livekit-agents
livekit-agents[google]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
google-generativeai
fastapi
uvicorn[standard]
python-dotenv
livekit>=1.0.12
livekit-plugins-noise-cancellation>=0.2.5
livekit-agents
livekit-plugins-google
google-generativeai
python-dotenv
langchain
duckduckgo-search
